+++ Cargo.toml
@@ -15,15 +15,13 @@ crate-type = ["cdylib", "lib", "staticlib"]
 [dependencies]
 rand = "0.4"
 bit-vec = "0.6"
-futures = "0.1"
 cfg-if = "0.1"
 
 #pairing = {package = "pairing_ce", path = "../pairing" }
 pairing = {package = "pairing_ce", version = "0.19" }
 byteorder = "1"
 
-futures-cpupool = {version = "0.1", optional = true}
-futures_new = {package = "futures", version = "0.3", default_features = false, features = ["executor"]}
+futures= {package = "futures", version = "0.3", default_features = false, features = ["executor"]}
 num_cpus = "1"
 crossbeam = {version = "0.7", optional = true}
 
@@ -35,11 +33,10 @@ tiny-keccak = {version = "1.4.2", optional = true}
 blake2-rfc = {version = "0.2.18", optional = true}
 
 [features]
-default = []
-# default = ["multicore"]
-#default = ["multicore", "nightly"]
+# default = []
+default = ["multicore"]
 #default = ["wasm"]
-multicore = ["futures-cpupool", "crossbeam", "futures_new/thread-pool"]
+multicore = ["crossbeam", "futures/thread-pool"]
 sonic = ["tiny-keccak", "blake2-rfc"]
 gm17 = []
 wasm = ["web-sys"]
