+++ src/groth16/generator.rs
@@ -35,7 +35,7 @@ use crate::domain::{
     Scalar
 };
 
-use crate::worker_new::{
+use crate::worker::{
     Worker
 };
