+++ src/groth16/verifier.rs
@@ -62,6 +62,6 @@ pub fn verify_proof<'a, E: Engine>(
             (&proof.a.prepare(), &proof.b.prepare()),
             (&acc.into_affine().prepare(), &pvk.neg_gamma_g2),
             (&proof.c.prepare(), &pvk.neg_delta_g2)
-        ].into_iter())
+        ].iter())
     ).unwrap() == pvk.alpha_g1_beta_g2)
 }
