+++ src/multicore.rs
@@ -6,27 +6,35 @@
 
 extern crate num_cpus;
 extern crate futures;
-extern crate futures_cpupool;
 extern crate crossbeam;
 
-pub(crate) use self::futures::{Future, IntoFuture, Poll};
-use self::futures_cpupool::{CpuPool, CpuFuture};
+use std::future::{Future};
+use std::task::{Context, Poll};
+use std::pin::{Pin};
+
 use self::crossbeam::thread::{Scope};
+
+use self::futures::future::{lazy};
+use self::futures::channel::oneshot::{channel, Sender, Receiver};
+use self::futures::executor::{block_on};
+use self::futures::executor::{ThreadPool};
 
 #[derive(Clone)]
 pub struct Worker {
     cpus: usize,
-    pool: CpuPool
+    pool: ThreadPool
 }
+
 
 impl Worker {
     // We don't expose this outside the library so that
     // all `Worker` instances have the same number of
     // CPUs configured.
+    
     pub(crate) fn new_with_cpus(cpus: usize) -> Worker {
         Worker {
             cpus: cpus,
-            pool: CpuPool::new(cpus)
+            pool: ThreadPool::builder().pool_size(cpus).create().expect("should create a thread pool for futures execution"),
         }
     }
 
@@ -38,18 +46,29 @@ impl Worker {
         log2_floor(self.cpus)
     }
 
-    pub fn compute<F, R>(
+    pub fn compute<F, T, E>(
         &self, f: F
-    ) -> WorkerFuture<R::Item, R::Error>
-        where F: FnOnce() -> R + Send + 'static,
-              R: IntoFuture + 'static,
-              R::Future: Send + 'static,
-              R::Item: Send + 'static,
-              R::Error: Send + 'static
+    ) -> WorkerFuture<T, E>
+        where F: FnOnce() -> Result<T, E> + Send + 'static,
+              T: Send + 'static,
+              E: Send + 'static
     {
-        WorkerFuture {
-            future: self.pool.spawn_fn(f)
-        }
+        let (sender, receiver) = channel();
+        let lazy_future = lazy(move |_| {
+            let res = f(); 
+
+            if !sender.is_canceled() {
+                let _ = sender.send(res);
+            }
+        });
+
+        let worker_future = WorkerFuture {
+            receiver
+        };
+
+        self.pool.spawn_ok(lazy_future);
+
+        worker_future
     }
 
     pub fn scope<'a, F, R>(
@@ -72,16 +91,33 @@ impl Worker {
 }
 
 pub struct WorkerFuture<T, E> {
-    future: CpuFuture<T, E>
+    receiver: Receiver<Result<T, E>>
 }
 
 impl<T: Send + 'static, E: Send + 'static> Future for WorkerFuture<T, E> {
-    type Item = T;
-    type Error = E;
+    type Output = Result<T, E>;
 
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error>
+    fn poll(self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output>
     {
-        self.future.poll()
+        let rec = unsafe { self.map_unchecked_mut(|s| &mut s.receiver) };
+        match rec.poll(cx) {
+            Poll::Ready(v) => {
+                if let Ok(v) = v {
+                    return Poll::Ready(v)
+                } else {
+                    panic!("Worker future can not have canceled sender");
+                }
+            },
+            Poll::Pending => {
+                return Poll::Pending;
+            }
+        }
+    }
+}
+
+impl<T: Send + 'static, E: Send + 'static> WorkerFuture<T, E> {
+    pub fn wait(self) -> <Self as Future>::Output {
+        block_on(self)
     }
 }
 
@@ -107,4 +143,34 @@ fn test_log2_floor() {
     assert_eq!(log2_floor(6), 2);
     assert_eq!(log2_floor(7), 2);
     assert_eq!(log2_floor(8), 3);
+}
+
+#[test]
+fn test_trivial_spawning() {
+    use self::futures::executor::block_on;
+
+    fn long_fn() -> Result<usize, ()> {
+        let mut i: usize = 1;
+        println!("Start calculating long task");
+        for _ in 0..1000000 {
+            i = i.wrapping_mul(42);
+        }
+
+        println!("Done calculating long task");
+
+        Ok(i)
+    }
+
+    let worker = Worker::new();
+    println!("Spawning");
+    let fut = worker.compute(|| long_fn());
+    println!("Done spawning");
+
+    println!("Will sleep now");
+
+    std::thread::sleep(std::time::Duration::from_millis(10000));
+
+    println!("Done sleeping");
+
+    let _ = block_on(fut);
 }
